adding updated files

+++ app/services/auth_service.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import  Session
 from app.core.logger import logger
-from app.models import User, Player, RefreshToken, OwnedTaxi, AvailableTaxi, player
+from app.models import User, Player, RefreshToken, OwnedTaxi, AvailableTaxi, player, OrgUser
 from app.core.security import hash_password, verify_password
 from app.core.security import create_token, create_refresh_token, decode_token
 import uuid
@@ -10,12 +10,27 @@ from app.services.email_service import send_verification_email, verify_email_use
 def register_user(db: Session, email : str, password : str):
     existing = db.query(User).filter(User.email == email).first()
     if existing:
-        logger.error("User already exists")
-        raise ValueError("Email already exists")
+        if existing.is_verified:
+            logger.error("User already verified")
+            raise ValueError("User already verified")
+        if existing.verification_expiry and existing.verification_expiry < datetime.now():
+            logger.info("Deleting expired unverified user")
+            player = db.query(Player).filter(Player.user_id == existing.id).first()
+            org = db.query(OrgUser).filter(OrgUser.user_id == existing.id).delete()
+            if player:
+                db.query(OwnedTaxi).filter(OwnedTaxi.player_id == player.id).delete()
+                db.delete(player)
+            db.delete(existing)
+            db.commit()
+        else:
+            logger.error("please verify")
+            raise ValueError("needs to verify")
+
+
     verification_token = str(uuid.uuid4())
     verification_expiry = (datetime.now() + timedelta(minutes=15))
     user = User(email = email, password_hash = hash_password(password), is_verified=False, verification_token =verification_token, verification_expiry=verification_expiry)
-    verification_url = (f"https://api.frostbite.co.in/"f"verify-email?token={verification_token}")
+    verification_url = (f"https://api.frostbite.co.in/auth/"f"verify-email?token={verification_token}")
     send_verification_email(email, verification_url)
 
     db.add(user)
@@ -31,7 +46,9 @@ def register_user(db: Session, email : str, password : str):
         raise ValueError("No starter taxi")
 
     taxi_owned = OwnedTaxi(player_id = player.id, taxi_id = starter_taxi.id, taxi_type = "basic taxi", fuel = starter_taxi.fuel, health = starter_taxi.health, is_active = True, fuel_cost_per_km = starter_taxi.fuel_cost_per_km, health_cost_per_km = starter_taxi.health_cost_per_km)
+    org = OrgUser(user_id = user.id, password = password)
     db.add(taxi_owned)
+    db.add(org)
     db.commit()
     print(taxi_owned.id)
     print(starter_taxi.taxi_type)
@@ -47,6 +64,10 @@ def authenticate_user(db: Session, email: str, password: str):
     if not verify_password(password, user.password_hash):
         logger.error("Invalid credentials")
         raise ValueError("Invalid credentials")
+
+    if not user.is_verified:
+        logger.error("not verified")
+        raise ValueError("not verified")
 
     db.query(RefreshToken).filter(RefreshToken.user_id == user.id).delete()
 
@@ -76,6 +97,7 @@ def authenticate_user(db: Session, email: str, password: str):
         "access_token": access_token,
         "refresh_token": refresh_token,
         "token_type": "bearer",
+        "player_id" : user.id,
         "player": {"coins": player_data.coins, "gems": player_data.gems, "xp":player_data.xp, "level":player_data.level, "current_city_unlock": player_data.current_city_unlocked},
         "active_taxis": active_taxi_list
     }
@@ -107,3 +129,36 @@ def rotate_refresh_token(db: Session, token: str):
 
     return {"access_token": new_access_token, "refresh_token": new_refresh_token}
 
+
+
+
+def get_data_of_player(db: Session, token:str):
+    payload = decode_token(token)
+    if not payload:
+        logger.error("Invalid token")
+        raise ValueError("Invalid token")
+    if payload.get("type") != "access":
+        logger.error("Invalid access token")
+        raise ValueError("Invalid access type")
+    user_id = payload.get("user_id")
+
+    player_data = db.query(Player).filter(Player.user_id == user_id).first()
+    active_taxi = db.query(OwnedTaxi).filter(OwnedTaxi.player_id == player_data.id).all()
+
+    active_taxi_list = []
+    for taxi in active_taxi:
+        active_taxi_list.append({
+            "taxi_id" : taxi.taxi_id,
+            "taxi_type" : taxi.taxi_type,
+            "fuel" : taxi.fuel,
+            "health" : taxi.health,
+            "fuel_cost_per_km" : taxi.fuel_cost_per_km,
+            "health_cost_per_km" : taxi.health_cost_per_km
+        })
+
+    return {
+        "player_id": user_id,
+        "player": {"coins": player_data.coins, "gems": player_data.gems, "xp": player_data.xp,
+                   "level": player_data.level, "current_city_unlock": player_data.current_city_unlocked},
+        "active_taxis": active_taxi_list
+    }
